Move relative_path_to onto AnchoredSystemPathBuf
