+++ crates/turborepo-paths/src/absolute_system_path.rs
@@ -9,7 +9,7 @@ use std::{
     fmt, fs,
     fs::Metadata,
     io,
-    path::{Component, Components, Path, PathBuf},
+    path::{Components, Path, PathBuf},
 };
 
 use path_clean::PathClean;
@@ -193,45 +193,6 @@ impl AbsoluteSystemPath {
     pub fn components(&self) -> Components<'_> {
         self.0.components()
     }
-
-    // Produces a path from self to other, which may include directory traversal
-    // tokens. Given that both parameters are absolute, we _should_ always be
-    // able to produce such a path. The exception is when crossing drive letters
-    // on Windows, where no such path is possible. Since a repository is
-    // expected to only reside on a single drive, this shouldn't be an issue.
-    pub fn relative_path_to(&self, other: &Self) -> AnchoredSystemPathBuf {
-        // Filter the implicit "RootDir" component that exists for unix paths.
-        // For windows paths, we may want an assertion that we aren't crossing drives
-        let these_components = self
-            .components()
-            .skip_while(|c| *c == Component::RootDir)
-            .collect::<Vec<_>>();
-        let other_components = other
-            .components()
-            .skip_while(|c| *c == Component::RootDir)
-            .collect::<Vec<_>>();
-        let prefix_len = these_components
-            .iter()
-            .zip(other_components.iter())
-            .take_while(|(a, b)| a == b)
-            .count();
-        #[cfg(windows)]
-        debug_assert!(
-            prefix_len >= 1,
-            "Cannot traverse drives between {} and {}",
-            self,
-            other
-        );
-
-        let traverse_count = these_components.len() - prefix_len;
-        // For every remaining non-matching segment in self, add a directory traversal
-        // Then, add every non-matching segment from other
-        let path = std::iter::repeat(Component::ParentDir)
-            .take(traverse_count)
-            .chain(other_components.into_iter().skip(prefix_len))
-            .collect::<PathBuf>();
-        AnchoredSystemPathBuf(path)
-    }
 }
 
 #[cfg(test)]
@@ -239,45 +200,6 @@ mod tests {
     use anyhow::Result;
 
     use super::*;
-
-    #[test]
-    fn test_relative_path_to() {
-        #[cfg(unix)]
-        let root_token = "/";
-        #[cfg(windows)]
-        let root_token = "C:\\";
-
-        let root = AbsoluteSystemPathBuf::new(
-            [root_token, "a", "b", "c"].join(std::path::MAIN_SEPARATOR_STR),
-        )
-        .unwrap();
-
-        // /a/b/c
-        // vs
-        // /a -> ../..
-        // /a/b/d -> ../d
-        // /a/b/c/d -> d
-        // /e/f -> ../../../e/f
-        // / -> ../../..
-        let test_cases: &[(&[&str], &[&str])] = &[
-            (&["a"], &["..", ".."]),
-            (&["a", "b", "d"], &["..", "d"]),
-            (&["a", "b", "c", "d"], &["d"]),
-            (&["e", "f"], &["..", "..", "..", "e", "f"]),
-            (&[], &["..", "..", ".."]),
-        ];
-        for (input, expected) in test_cases {
-            let mut parts = vec![root_token];
-            parts.extend_from_slice(input);
-            let target =
-                AbsoluteSystemPathBuf::new(parts.join(std::path::MAIN_SEPARATOR_STR)).unwrap();
-            let expected =
-                AnchoredSystemPathBuf::from_raw(expected.join(std::path::MAIN_SEPARATOR_STR))
-                    .unwrap();
-            let result = root.relative_path_to(&target);
-            assert_eq!(result, expected);
-        }
-    }
 
     #[test]
     fn test_create_absolute_path() -> Result<()> {
